backend/requirements.txt
fastapi
uvicorn
requests
joblib
numpy
pandas
scikit-learn
openai-whisper
yt-dlp